requirements.txt
# =============================================================================
# DEPRECATION WARNING:
#
# The file `requirements.txt` does not influence the package dependencies and
# will not be automatically created in the next version of PyScaffold (v4.x).
#
# Please have look at the docs for better alternatives
# (`Dependency Management` section).
# =============================================================================
eventlet~=0.33
celery~=5.3
cryptography~=41.0
gevent~=23.7
gunicorn~=19.9
falcon~=3.1
jsontas~=1.3
packageurl-python~=0.11
etcd3gw~=2.3
etos_lib==3.2.2
